+++ packages/orq-rc/src/models/operations/createdatasource.ts
@@ -751,7 +751,7 @@ export const CreateDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01K855G3ZJKD15Y1XA9G789QZD"),
+  _id: z.string().default("01K85KQYYHTWTSSGQJFCK8SYR3"),
   display_name: z.string(),
   description: z.string().optional(),
   status: CreateDatasourceStatus$inboundSchema,
@@ -795,7 +795,7 @@ export const CreateDatasourceResponseBody$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   CreateDatasourceResponseBody
 > = z.object({
-  id: z.string().default("01K855G3ZJKD15Y1XA9G789QZD"),
+  id: z.string().default("01K85KQYYHTWTSSGQJFCK8SYR3"),
   displayName: z.string(),
   description: z.string().optional(),
   status: CreateDatasourceStatus$outboundSchema,